community-report-system/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.0.1
email-validator==2.1.0
sqlalchemy==2.0.23